Maya aov fixes

+++ Prism/Plugins/Apps/Maya/Scripts/Prism_Maya_Functions.py
@@ -1042,7 +1042,8 @@ class Prism_Maya_Functions(object):
 
 			aAovs = maovs.AOVInterface().getAOVNodes(names=True)
 			aAovs = [x for x in aAovs if cmds.getAttr(x[1] + '.enabled')]
-			if cmds.getAttr("defaultArnoldRenderOptions.aovMode") != 0 and len(aAovs) > 0:
+			multichannel = cmds.getAttr("defaultArnoldDriver.mergeAOVs") == 1 
+			if cmds.getAttr("defaultArnoldRenderOptions.aovMode") != 0 and not multichannel and len(aAovs) > 0:
 				outputPrefix = "../" + outputPrefix
 			#	if len(rlayers) > 1:
 			#		outputPrefix = "../" + outputPrefix
@@ -1077,13 +1078,15 @@ class Prism_Maya_Functions(object):
 			rSettings["vr_sepRGBA"] = cmds.getAttr("vraySettings.relements_separateRGBA")
 			rSettings["vr_animation"] = cmds.getAttr("vraySettings.animType")
 
-			cmds.setAttr("vraySettings.imageFormatStr", "exr", type="string")
+			multichannel = cmds.getAttr("vraySettings.imageFormatStr") not in ["exr (multichannel)", "exr (deep)"]
+			if not multichannel:
+				cmds.setAttr("vraySettings.imageFormatStr", "exr", type="string")
 			cmds.setAttr("vraySettings.animType", 1)
 
 			aovs = cmds.ls(type='VRayRenderElement')
 			aovs = [x for x in aovs if cmds.getAttr(x + '.enabled')]
 
-			if cmds.getAttr("vraySettings.relements_enableall") != 0 and len(aovs) > 0:
+			if cmds.getAttr("vraySettings.relements_enableall") != 0 and not multichannel and len(aovs) > 0:
 				try:
 					shutil.rmtree(os.path.dirname(rSettings["outputName"]))
 				except:
@@ -1114,7 +1117,11 @@ class Prism_Maya_Functions(object):
 
 			aovs = cmds.ls(type='RedshiftAOV')
 			aovs = [[cmds.getAttr(x + ".name"), x] for x in aovs if cmds.getAttr(x + '.enabled')]
+			for aov in aovs:
+				if cmds.getAttr(aov[1] + ".aovType") == "Beauty":
+					rSettings["outputName"] = rSettings["outputName"].replace("beauty", aov[0])
 
+			#multichannel = cmds.getAttr("redshiftOptions.exrForceMultilayer") == 1
 			if cmds.getAttr("redshiftOptions.aovGlobalEnableMode") != 0 and len(aovs) > 0:
 				for i in aovs:
 					cmds.setAttr(i[1] + ".filePrefix", "<BeautyPath>/../<RenderPass>/%s" % os.path.basename(outputPrefix).replace("beauty", i[0]), type="string")
@@ -1474,57 +1481,72 @@ class Prism_Maya_Functions(object):
 
 			validNodes = [ x for x in origin.nodes if self.isNodeValid(origin, x)]
 			if not update or len(validNodes) == 0:
-				refDlg = QDialog()
-
-				refDlg.setWindowTitle("Create Reference")
-				rb_reference = QRadioButton("Create reference")
-				rb_reference.setChecked(True)
-				rb_import = QRadioButton("Import objects only")
-				rb_applyCache = QRadioButton("Apply as cache to selected objects")
-				w_namespace = QWidget()
-				nLayout = QHBoxLayout()
-				nLayout.setContentsMargins(0,15,0,0)
-				chb_namespace = QCheckBox("Create namespace")
-				chb_namespace.setChecked(True)
-				e_namespace = QLineEdit()
-				e_namespace.setText(fileName[0])
-				nLayout.addWidget(chb_namespace)
-				nLayout.addWidget(e_namespace)
-				chb_namespace.toggled.connect(lambda x: e_namespace.setEnabled(x))
-				w_namespace.setLayout(nLayout)
-
-				rb_applyCache.toggled.connect(lambda x: w_namespace.setEnabled(not x))
-				if fileName[1] != ".abc" or len(cmds.ls(selection=True)) == 0:
-					rb_applyCache.setEnabled(False)
+				# default settings
+				mode = "reference"
+				useNamespace = True
+				namespace = fileName[0]
 				
-				bb_warn = QDialogButtonBox()
-				bb_warn.addButton("Ok", QDialogButtonBox.AcceptRole)
-				bb_warn.addButton("Cancel", QDialogButtonBox.RejectRole)
+				if self.core.uiAvailable:
+					refDlg = QDialog()
 
-				bb_warn.accepted.connect(refDlg.accept)
-				bb_warn.rejected.connect(refDlg.reject)
+					refDlg.setWindowTitle("Create Reference")
+					rb_reference = QRadioButton("Create reference")
+					rb_reference.setChecked(mode=="reference")
+					rb_import = QRadioButton("Import objects only")
+					rb_reference.setChecked(mode=="import")
+					rb_applyCache = QRadioButton("Apply as cache to selected objects")
+					rb_reference.setChecked(mode=="applyCache")
+					w_namespace = QWidget()
+					nLayout = QHBoxLayout()
+					nLayout.setContentsMargins(0,15,0,0)
+					chb_namespace = QCheckBox("Create namespace")
+					chb_namespace.setChecked(useNamespace)
+					e_namespace = QLineEdit()
+					e_namespace.setText(namespace)
+					nLayout.addWidget(chb_namespace)
+					nLayout.addWidget(e_namespace)
+					chb_namespace.toggled.connect(lambda x: e_namespace.setEnabled(x))
+					w_namespace.setLayout(nLayout)
 
-				bLayout = QVBoxLayout()
-				bLayout.addWidget(rb_reference)
-				bLayout.addWidget(rb_import)
-				bLayout.addWidget(rb_applyCache)
-				bLayout.addWidget(w_namespace)
-				bLayout.addWidget(bb_warn)
-				refDlg.setLayout(bLayout)
-				refDlg.setParent(self.core.messageParent, Qt.Window)
-				refDlg.resize(400,100)
+					rb_applyCache.toggled.connect(lambda x: w_namespace.setEnabled(not x))
+					if fileName[1] != ".abc" or len(cmds.ls(selection=True)) == 0:
+						rb_applyCache.setEnabled(False)
+					
+					bb_warn = QDialogButtonBox()
+					bb_warn.addButton("Ok", QDialogButtonBox.AcceptRole)
+					bb_warn.addButton("Cancel", QDialogButtonBox.RejectRole)
 
-				action = refDlg.exec_()
+					bb_warn.accepted.connect(refDlg.accept)
+					bb_warn.rejected.connect(refDlg.reject)
 
-				if action == 0:
-					doRef = False
-					importOnly = False
-					applyCache = False
+					bLayout = QVBoxLayout()
+					bLayout.addWidget(rb_reference)
+					bLayout.addWidget(rb_import)
+					bLayout.addWidget(rb_applyCache)
+					bLayout.addWidget(w_namespace)
+					bLayout.addWidget(bb_warn)
+					refDlg.setLayout(bLayout)
+					refDlg.setParent(self.core.messageParent, Qt.Window)
+					refDlg.resize(400,100)
+
+					action = refDlg.exec_()
+
+					if action == 0:
+						doRef = False
+						importOnly = False
+						applyCache = False
+					else:
+						doRef = rb_reference.isChecked()
+						applyCache = rb_applyCache.isChecked()
+						if chb_namespace.isChecked():
+							nSpace = e_namespace.text()
+						else:
+							nSpace = ":"
 				else:
-					doRef = rb_reference.isChecked()
-					applyCache = rb_applyCache.isChecked()
-					if chb_namespace.isChecked():
-						nSpace = e_namespace.text()
+					doRef = mode == "reference"
+					applyCache = mode == "applyCache"
+					if useNamespace:
+						nSpace = namespace
 					else:
 						nSpace = ":"
 			else:
@@ -1611,7 +1633,8 @@ class Prism_Maya_Functions(object):
 					importedNodes = cmds.file(impFileName, i=True, returnNewNodes=True)
 				except Exception as e:
 					importedNodes = []
-					QMessageBox.warning(self.core.messageParent, "Import error", "An error occured while importing the file:\n\n%s\n\n%s" % (impFileName, str(e)))
+					msg = "An error occured while importing the file:\n\n%s\n\n%s" % (impFileName, str(e))
+					self.core.popup(msg, title="Import error")
 
 		for i in importedNodes:
 			cams = cmds.listCameras()
